no size_t, fixes for darwin
git-svn-id: https://bugengine.googlecode.com/svn/trunk@439 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/engine/rtti/api/rtti/engine/methodinfo.script.hh
@@ -25,9 +25,9 @@ published:
     TypeInfo const                  returnType;
     minitl::vector<const ParamInfo> params;
 public:
-    mutable Value (*call)(Value* params, size_t nparams);
+    mutable Value (*call)(Value* params, u32 nparams);
 
-    OverloadInfo(const TypeInfo& returnType, Value (*call)(Value* params, size_t nparams))
+    OverloadInfo(const TypeInfo& returnType, Value (*call)(Value* params, u32 nparams))
         :   returnType(returnType)
         ,   params(rttiArena())
         ,   call(call)
@@ -40,7 +40,7 @@ published:
     minitl::vector<OverloadInfo>    overloads;
     OverloadInfo const              vararg;
 published:
-    Value operator()(Value* params, size_t nparams) const;
+    Value operator()(Value* params, u32 nparams) const;
 public:
     MethodInfo();
 };
